Improved encoding detection a bit
The Windows 1252 encoding is only available in .NET Framework, so using the "GetEncodingBySeek" method would fail in some scenarios. In .NET, the Windows 1252 encoding seems to have been replaced with ISO 8859-1, so the method is now also checking against that encoding.

To my knowledge there isn't a TryGet approach to getting an encoding, so for now, each attempt to load an encoding is wrapped with a try/catch statement.